+++ futurehome2mqtt/pyfimptoha/client.py
@@ -1,8 +1,9 @@
 import json
+import time
 
 import pyfimptoha.fimp as fimp
 import pyfimptoha.homeassistant as homeassistant
-from pyfimptoha.mode import Mode
+import pyfimptoha.mode as mode
 
 
 class Client:
@@ -27,20 +28,17 @@ class Client:
             self._mqtt = mqtt
             mqtt.on_message = self.on_message
 
-        # Add Modus sensor (home, sleep, away and vacation)
-        # todo Find a way to auto discover the value. Sensor value is currently
-        # empty until changed by the system/user
         self.start()
 
     def start(self):
         mqtt = self._mqtt
 
-        # todo refactor Mode
-        # mode = Mode()
-        # message = mode.get_component()
-        # self.publish_messages([message])
+        # Request FIMP mode
+        topic_receive_mode = "pt:j1/mt:rsp/rt:app/rn:homeassistant/ad:mode"
+        mqtt.subscribe(topic_receive_mode)
+        fimp.send_mode_request(mqtt)
 
-        # Send FIMP discover request
+        # Request FIMP devices
         topic_discover = "pt:j1/mt:rsp/rt:app/rn:homeassistant/ad:flow1"
         mqtt.subscribe(topic_discover)
         fimp.send_discovery_request(mqtt)
@@ -62,6 +60,13 @@ class Client:
                 mqtt=self._mqtt,
                 selected_devices=self._selected_devices,
             )
+        elif msg.topic == "pt:j1/mt:rsp/rt:app/rn:homeassistant/ad:mode":
+            # Create mode sensor  (home, away, sleep and vacation)
+            data = json.loads(payload)
+            mode.create(
+                mqtt=self._mqtt,
+                data=data,
+            )
         elif msg.topic == "homeassistant/status" and payload == "online":
-            print(msg.topic)
+            # Home Assistant was restarted - Push everything again
             self.start()